Moved some methods from the codebehind in SearchView to the viewmodel
