+++ HolidayMakerClient/HolidayMakerClient/ViewModel/SearchViewModel.cs
@@ -4,8 +4,12 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Media;
 
 namespace HolidayMakerClient.ViewModel
 {
@@ -36,6 +40,7 @@ namespace HolidayMakerClient.ViewModel
         public bool SortingAttributes { get; set; }
         public ObservableCollection<Home> HomeList { get; }
         public ObservableCollection<Home> SortedHomeList { get; }
+        public List<FontIcon> FontIconList = new List<FontIcon>();
         #endregion
 
         #region Methods
@@ -99,6 +104,7 @@ namespace HolidayMakerClient.ViewModel
         {
             SortedHomeList.Clear();
             IOrderedEnumerable<Home> orderedList;
+
             switch (parameter)
             {
                 case "bttn_SortLocation":
@@ -127,6 +133,63 @@ namespace HolidayMakerClient.ViewModel
                         SortedHomeList.Add(x);
                     break;
             }
+
+        }
+        
+        public void SortColumns(Button sortButton)
+        {
+            //Find the child of type FontIcon to the button
+            FontIcon glyph = FindFontIconChild((DependencyObject)sortButton);
+
+            //Fill the glyph property of the fonticon. Either an up-arrow or down arrow
+            //and set if the sorting is ascending or descending
+            if ((glyph).Glyph == "\uE96E")
+            {
+                (glyph).Glyph = "\uE96D";
+                IsAscending = false;
+            }
+            else
+            {
+                (glyph).Glyph = "\uE96E";
+                IsAscending = true;
+            }
+
+            //Clear the arrow from any button except the clicked one
+            ClearGlyphs(glyph);
+            //Sort the list
+            SortList(((FrameworkElement)sortButton).Name);
+        }
+
+        private void ClearGlyphs(FontIcon fontIcon)
+        {
+            //Go through the list of fonticons. if the fonticon is the one in the parameter list.
+            //do nothing with it.
+            foreach(var fi in FontIconList)
+                if (fontIcon != fi)
+                    fi.Glyph = "";
+        }
+        private FontIcon FindFontIconChild(DependencyObject parent)
+        {
+            //If the font icon is found then we just exit the method with the object in hand.
+            if (parent is FontIcon)
+                return (FontIcon)parent;
+
+            //create a fonticon variable.
+            FontIcon foundFontIcon = null;
+
+            //go through the children of the visual element and see if one of them is of the type FontIcon
+            for (int i = 0; i < VisualTreeHelper.GetChildrenCount(parent); i++)
+            {
+                var child = VisualTreeHelper.GetChild(parent, i);
+
+                if (child is FontIcon)
+                    foundFontIcon = (FontIcon)child;
+                else
+                    //If the icon isn't found. Recurse through this method again
+                    foundFontIcon = FindFontIconChild(child);
+            }
+            //return with the fonticon
+            return foundFontIcon;
         }
         #endregion
     }
